requirements.txt
django==4.2.4
djangorestframework==3.14.0
scikit-learn==1.2.2
joblib==1.3.2
pandas==1.5.3
numpy==1.23.5
keras==2.13.1
tensorflow==2.13
imbalanced-learn==0.11.0